Abstract out content manager / angular adapter into microservices helper package and decrease footprint of microservices adaptor code in main application
