+++ modules/blue/css/blueInnerStyles.css
@@ -1,4 +1,4 @@
 .inner{
     font-size: 21px;
-    color: #770303;
+    color: red;
 }
