+++ modules/microservices-helper/index.js
@@ -1,4 +1,4 @@
-export { getModuleOptions, getComponents, registerModule } from './helpers/registerModuleHelper';
+export { registerModule } from './helpers/registerModuleHelper';
 export { connect } from './helpers/connectWrapper';
 export { store, store$, setStore } from './helpers/storeHelper';
-export { getParams } from './utils/paramsUtil';
+export { initializeModules } from './helpers/angularAdapter';
