+++ src/js/modules/appModule.js
@@ -3,39 +3,37 @@ import '@uirouter/angularjs';
 import './basicModule';
 import '../modules/content';
 import { RootApp } from '../components/rootApp';
-import { ContentMananger } from '../services/contentManager';
 import { TriggerHandlerService } from '../services/triggerHanderService';
+
+import reducerRegistry from '../../store/reducers/reducerRegistery';
+import { store } from '../../store/stateStore';
+import { epic$ } from '../../store/epics/epicRegistery';
+import { combineEpics } from 'redux-observable';
+
+import { initializeModules } from 'microservices-helper';
 
 // -----------------------
 import 'blue';
 import 'green';
 // -----------------------
 
-import { getComponents, getModuleOptions } from 'microservices-helper';
-
-const moduleComponents = getComponents();
-const moduleOptions = getModuleOptions();
-
-let $stateProviderSaved;
-
 const app = angular.module('app', ['ui.router', 'app.basic', 'app.red', 'app.yellow']);
 
 app.config(['$stateProvider', ($stateProvider) => {
-    $stateProviderSaved = $stateProvider;
+    console.log('Original config callback');
 }]);
 
 app.run(['contentManager', (contentMananger) => {
-    contentMananger.initialize($stateProviderSaved);
-
-    moduleOptions.forEach((options) => {
-        contentMananger.registerContent(options);
-    });
+    console.log('Original run callback');
 }]);
 
 app.component('rootApp', RootApp);
-app.service('contentManager', ContentMananger);
-app.service('triggerHandler', TriggerHandlerService);
 
-moduleComponents.forEach((component) => {
-    app.component(component.name, component.value);
+const passedServices = ['experimentalService'];
+
+initializeModules(app, TriggerHandlerService, passedServices, {
+    store,
+    epic$,
+    reducerRegistry,
+    combineEpics,
 });
